imagetotext1/requirements.txt
gunicorn==20.1.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
pytesseract==0.3.10
Flask==2.2.3
opencv-contrib-python==4.7.0.72
Flask-RESTful==0.3.9
Werkzeug~=2.2.3